pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "openai>=1.82.0",
    "passlib>=1.7.4",
    "pdfplumber>=0.11.6",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "twilio>=9.6.1",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]
